+++ src/sim/mecafil.cc
@@ -142,7 +142,7 @@ void Mecafil::storeDirections()
      */
     
     // for the extremities, the direction of the nearby segment is used.
-    for ( unsigned d = 0; d < DIM; ++d )
+    for ( size_t d = 0; d < DIM; ++d )
     {
         rfDir[d]     = rfDiff[d];
         rfDir[d+end] = rfDiff[d+end-DIM];
@@ -682,7 +682,7 @@ void add_rigidityF(const size_t nbt, const real* X, const real R1, real* Y)
     real      * Z = Y + nbt;
     real const* E = X + nbt + DIM;
     #pragma ivdep
-    for ( unsigned d = 0; d < DIM; ++d )
+    for ( size_t d = 0; d < DIM; ++d )
     {
         Y[d+DIM] -= R1 * ((X+DIM)[d]+(X+DIM*3)[d]) + R4 * ((X+DIM)[d]-(X+DIM*2)[d]) - R2 * X[d];
         Z[d    ] -= R1 * ((E-DIM)[d]+(E-DIM*3)[d]) + R4 * ((E-DIM)[d]-(E-DIM*2)[d]) - R2 * E[d];
@@ -697,7 +697,7 @@ void add_rigidityF(const size_t nbt, const real* X, const real R1, real* Y)
  */
 void add_rigidity(size_t A, size_t B, size_t C, const real* X, const real R1, real* Y)
 {
-    for ( unsigned d = 0; d < DIM; ++ d )
+    for ( size_t d = 0; d < DIM; ++ d )
     {
         real x = 2*X[B*DIM+d] - ( X[A*DIM+d] + X[C*DIM+d] );
         Y[A*DIM+d] += x * R1;
